backend/requirements.txt
# Flask & Web Framework
Flask==3.1.2
flask-cors==6.0.2
Werkzeug==3.1.2

# Supabase & Database
supabase==2.28.0
postgrest-py==0.15.4
realtime-py==1.0.0

# Authentication & Security
PyJWT==2.10.0
python-dotenv==1.2.1
cryptography==46.0.5

# Data Validation
pydantic==2.12.5
pydantic[email]==2.12.5
email-validator==2.1.0

# HTTP & Requests
requests==2.32.3
httpx==0.27.0

# Production Server
gunicorn==25.1.0
python-multipart==0.0.6

# Optional: Email & Notifications
# sendgrid==6.12.0
# twilio==9.0.0

# Development & Testing (Opcional)
# pytest==8.0.0
# pytest-flask==1.3.0
# factory-boy==3.3.0

# Utilities
python-dateutil==2.8.2
click==8.1.7
